Condition added: Must start with "!"
Added condition on main.py (Line 38 - 52) where bot checks if user starts their input with "!" before it sends a message to the server. This is to prevent from activating the bot by accident.

Line 43: user_message is where the message content is stored in string format.

Line 46: the message is checked to see if it starts with "!". If it does not start with ! the function stops immediately (return)

Line 50: Once ! is confirmed user_message removes the ! by setting the second letter (1) as the starting message which is the first letter of the first word if ! is included

Line 55: Logs message in the terminal