Added Pre-generated messages
Removed elif chain of the towns and it's recommended spots and replaced with Dictionary list for a cleaner code, allows new places to be added easily as well.

Added Reply_List which contains a list of Pre-generated list to be combined with the chosen spot for the user.

rework how message is returned in Line 57-63. Now the place is taken and is formatted with a random message from the Reply_List and then returned as message.

Minor Indent change on comment for main.py